Add files via upload
Following is the description of the files.
unemployment_analysis.csv --> is the dataset which is being used. Only 'Avg' column was added at the end to list the average of unemployment rate for all the countries.
Unemployment_Analyss.ipynb --> is the jupyter notebook, you can just export it and should be good to go.
Unemployment_Analysis.pdf --> is the pdf version of the jupyter notebook, it just for better reading experiance.